FX Luminaire Luxor low voltage integration
Support for the FX Luminaire low voltage controller.  See
https://www.fxl.com

The controller groups individual lights into Groups.  Entire Groups
are turned on and off together.  This integration exposes each Group
on the controller as a LightEntity.  Brightness for the group is
supported.

The controller can have several Themes, consisting of Groups, where each
group can have an associated brightness.  Themes can be turned on and
off.  This integration will optionally expose each Theme as a
LightEntity, in addition to the Groups.